requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
tensorflow==2.13.0
keras==2.13.0
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
Jinja2==3.1.2
